+++ tradingstrategy/candle.py
@@ -22,6 +22,7 @@ import pyarrow as pa
 from dataclasses_json import dataclass_json
 
 from tradingstrategy.chain import ChainId
+from tradingstrategy.timebucket import TimeBucket
 from tradingstrategy.types import UNIXTimestamp, USDollarAmount, BlockNumber, PrimaryKey, NonChecksummedAddress, \
     RawChainId
 from tradingstrategy.utils.groupeduniverse import PairGroupedUniverse
@@ -554,12 +555,15 @@ class GroupedCandleUniverse(PairGroupedUniverse):
         return GroupedCandleUniverse(df=Candle.to_qstrader_dataframe(), timestamp_column="Date", fix_wick_threshold=None)
 
     @staticmethod
-    def create_from_single_pair_dataframe(df: pd.DataFrame) -> "GroupedCandleUniverse":
+    def create_from_single_pair_dataframe(
+            df: pd.DataFrame,
+            bucket: TimeBucket | None = None,
+    ) -> "GroupedCandleUniverse":
         """Construct universe based on a single trading pair data.
 
         Useful for synthetic data/testing.
         """
-        return GroupedCandleUniverse(df)
+        return GroupedCandleUniverse(df, time_bucket=bucket)
 
     @staticmethod
     def create_from_multiple_candle_datafarames(dfs: Iterable[pd.DataFrame]) -> "GroupedCandleUniverse":